fix models

+++ internal/models/orders.go
@@ -1,9 +1,9 @@
 package models
 
 type Orders struct {
-	UserID     string `json:"-" db:"-"`
-	ID         int    `json:"number" db:"id"`
-	Status     string `json:"status" db:"status"`
-	Accrual    int    `json:"accrual" db:"accrual"`
-	UploadedAt string `json:"uploaded_at" db:"uploaded_at"`
+	UserID     string  `json:"-" db:"-"`
+	ID         int     `json:"number" db:"id"`
+	Status     string  `json:"status" db:"status"`
+	Accrual    float64 `json:"accrual" db:"accrual"`
+	UploadedAt string  `json:"uploaded_at" db:"uploaded_at"`
 }
